requirements.txt
playwright
python-telegram-bot==20.7
pandas
requests
aiohttp
